feat: Add @radix-ui/react-checkbox package
This commit adds the "@radix-ui/react-checkbox" package to the project's dependencies in the package.json file. This package will be used to implement checkboxes in the UI components.

Refactor AuthService register method

The AuthService register method in the auth.service.ts file has been refactored to accept an optional "lastName" parameter. This allows for flexibility when registering users, as the last name is no longer required.

Update authRouter in auth.ts

The authRouter in the auth.ts file has been updated to pass the "lastName" parameter as optional when calling the AuthService register method. This aligns with the changes made in the AuthService.

Update ZRegisterInput type

The ZRegisterInput type in the auth.ts file has been updated to make the "lastName" field optional. This ensures that the type definition matches the updated register method in the AuthService.

Remove ReactMic component from VoiceRecorder

The ReactMic component has been removed from the VoiceRecorder component in the voice-recorder.tsx file. This component was no longer being used and has been deleted.

Add Checkbox component

A new Checkbox component has been added to the project in the checkbox.tsx file. This component uses the "@radix-ui/react-checkbox" package to implement custom checkboxes in the UI.

Update RegisterMobileView and LoginMobileView

The RegisterMobileView and LoginMobileView components have been updated to include a "Remember me" checkbox. This checkbox is implemented using the newly added Checkbox component.

Remove unused code in ssg-init.ts

Unused code related to server-side rendering has been removed from the ssg-init.ts file. This code was commented out and no longer needed.

Update RegisterDekstopView

The RegisterDekstopView component has been updated to remove the "Or continue with" separator and the "Sign up with Google" button. These UI elements are no longer needed and have been deleted.